pyproject.toml
[project]
name = "aws-commons"
version = "0.1.0"
description = "Shared AWS infrastructure and CDK utilities"
authors = [
    { name = "Aaron Mamparo", email = "aaronmamparo@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13,<4"
package-mode = false
dependencies = [
    "boto3 (>=1.37.37,<2.0.0)",
    "aws-cdk-lib (>=2.190.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "injector (>=0.22.0,<0.23.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.6"
mypy = "^1.15.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

